// react-customer-engage-program/modules/migration-request-tracker/enums.ts
export enum NavigationTabs {
    ACTION = "Action",
    TIMELINE = "Timeline",
    MEMBERS = "Members",
    FILES = "Files",
    LEAVE_CONVERSATION = "Leave this Conversation",
    FULL_SCREEN = "Full Screen"
}

export enum MigrationProcess {
    SIGN_AN_NDA = "Sign an NDA",
}

export enum ProcessStatus {
    TODO = "ToDo",
    WIP = "WIP",
    COMPLETED = "Completed",
}

// DUMMY-DATA
export enum ConversationTypes {
    CHOSEN_MIGRATION_REQUEST = "chosenMigrationRequest",
    CHOSEN_MIGRATION_TYPE = "chosenMigrationType",
    TECHNOLOGIES = "technologies",
    PROCESS_INITIATED = "processInitiated",
    PROCESS_UPDATED = "Progress Update",
    FILE_UPLOAD = "fileupload",
    QUERY = "query",
    MEETING = "Meeting"
}